### 2022-03-17 #64 @lbeziaud
[¬Re]Simulating socioeconomic-based affirmative action
**Original article:** Reardon, Sean F., et al. "What Levels of Racial Diversity Can Be Achieved with Socioeconomic‐Based Affirmative Action? Evidence from a Simulation Model." Journal of Policy Analysis and Management 37.3 (2018): 630-657. https://doi.org/10.1002/pam.22056

**Scientific domain:** Education Policy
