### 2022-03-30 #64 @rougier
@lbeziaud Quick comment on the paper: you re-use the original figures which is good and ease the comparison. Problem is that we're not alllowed to do that because of copyright of the original paper (unless it is CC-BY). Could try to contact the journal to ask for permission to re-use the figures (explaining that the original article will be cited properly)??
