### 2022-03-30 #64 @lbeziaud
> Problem is that we're not alllowed to do that because of copyright of the original paper (unless it is CC-BY). Could try to contact the journal to ask for permission to re-use the figures (explaining that the original article will be cited properly)??

Thank you for noticing. I did not think about this issue and will contact the journal. I can produce a version without the original figures if necessary.

> Note that is a failed replication and we might need to contact original authors at some point if they want to give extra information.

Regarding the failed replication, only a subset of the model (however central to the conclusions of the original paper) was not reproduced successfully (due to our misunderstanding of some instructions). We originally thought about submitting as [~Re]. We tried to contact the first two authors in May 2021 for clarification but were left without an answer.
