### 2022-03-30 #64 @rougier
I would be best with original figures, but we can do without it if you don't obtain permission.
