+++ reardon.py
@@ -28,18 +28,40 @@ ach_var = np.sum(RACE_WEIGHT * (ACH_STD**2 + ACH_LOC**2 - ach_loc_**2))
 
 
 def run(
-    num_years=30, num_active=4, wgt_aa_race=0, wgt_aa_ses=0, wgt_recr_race=0, seed=None
+    num_years=30,
+    num_active=4,
+    wgt_aa_race=0,
+    wgt_aa_ses=0,
+    wgt_recr_race=0,
+    seed=None,
+    aa_formula1=False,
+    aa_penalize=False,
+    percept_qual_07=False,
 ):
+    """
+
+    :param num_years:
+    :param num_active: number of colleges using policies
+    :param wgt_aa_race: weight of race-based afformative action
+    :param wgt_aa_ses: weight of SES-based affirmative action
+    :param wgt_recr_race: weight of race-based recruiting
+    :param seed: seed for random number generator
+    :param aa_formula1: alternative SES-based AA
+    :param aa_penalize: alternative SES-based AA
+    :param percept_qual_07: alternative perception quality
+    :return:
+    """
     rng = np.random.default_rng(seed)
 
     qual = np.empty((num_years, NUM_COLLS))
     qual[0] = rng.normal(QUALITY_LOC, QUALITY_STD, size=NUM_COLLS)
-    qual[0, ::-1].sort()  # sort by (descending) qual
+    qual[0, ::-1].sort()  # sort by (descending) quality
 
+    # mask is true if college (top num_active) will use policy
     active = np.zeros(NUM_COLLS, dtype=bool)
     active[:num_active] = True
 
-    expyield = 0.2 + 0.6 * np.argsort(qual[0])[::-1] / (NUM_COLLS - 1)
+    expected_yield = 0.2 + 0.6 * np.argsort(qual[0])[::-1] / (NUM_COLLS - 1)
 
     race = rng.choice(4, size=(num_years, NUM_STUDS), p=RACE_WEIGHT)
     minority = (race == 1) | (race == 2)  # Black or Hispanic
@@ -51,16 +73,22 @@ def run(
     esd = np.sqrt(ACH_STD**2 - b**2 * RES_STD**2)
     ach = a[race] + b[race] * res + esd[race] * rng.normal(size=(num_years, NUM_STUDS))
 
-    sach_rel = np.clip(0.7 + 0.1 * res, 0.5, 0.9)
-    sach_err_std = np.sqrt(ach_var * (1 - sach_rel) / sach_rel)
-    sach_err = rng.normal(0, sach_err_std, (num_years, NUM_STUDS))
-    sach = ach + 0.1 * res + sach_err
+    stud_obs_ach_reliability = np.clip(0.7 + 0.1 * res, 0.5, 0.9)
+    stud_obs_ach_err_std = np.sqrt(
+        ach_var * (1 - stud_obs_ach_reliability) / stud_obs_ach_reliability
+    )
+    stud_obs_ach_err = rng.normal(0, stud_obs_ach_err_std, (num_years, NUM_STUDS))
+    stud_obs_ach = ach + 0.1 * res + stud_obs_ach_err
 
-    cach_err_std = np.sqrt(ach_var * (1 - 0.8) / 0.8)
-    cach_err = rng.normal(0, cach_err_std, (num_years, NUM_COLLS, NUM_STUDS))
-    cach = (ach + 0.1 * res)[:, np.newaxis, :] + cach_err
+    coll_obs_ach_err_std = np.sqrt(ach_var * (1 - 0.8) / 0.8)
+    coll_obs_ach_err = rng.normal(
[…]
