### 2022-03-31 #64 @lbeziaud
@rougier We contacted the editors and were redirected to Wiley which asks for $500 for reproduction rights. We removed the original figures from our paper.
