### 2022-03-31 #64 @rougier
That's unbelievable. If you want, you can leave an empty box in place of the original figure and write something like "Original figure cannot be reproduced here because Wiley ask $500 for reproduction rights".
