### 2022-03-31 #64 @lbeziaud
I was not expecting such an answer from the publisher... What I did was use the space to increase the size of our figures, added page numbers to the original figures, and a footnote explaining that we did not receive permission to include the figures (without mentioning the outrageous $500).
