### 2022-03-31 #64 @rougier
I found what seems to be a preprint at: https://cepa.stanford.edu/sites/default/files/wp15-04-v201712.pdf. I think you can reuse the figures from this one whch is prior to the publication. I'll check stanford policy on preprint but I think it should be ok.
