### 2022-03-31 #64 @rougier
Not clear, I'll ask them.

> In consideration for your agreement to the terms and conditions contained here, Stanford grants you a personal, non-exclusive, non-transferable license to access and use the Sites. User may download material from the Sites only for User’s own personal, non-commercial use. User may not otherwise copy, reproduce, retransmit, distribute, publish, commercially exploit or otherwise transfer any material. The burden of determining that use of any information, software or any other content on the Site is permissible rests with User.
