### 2022-03-31 #64 @lbeziaud
Thank you for the help investigating copyright issues! I added a link to this preprint in the footnote I was mentioning. The figures are actually black and white versions of the one published by Wiley.
