### 2022-03-31 #64 @lbeziaud
> I think you can include them.

The submission has been updated to include the figures from the preprint as you suggested, with updated references.
