### 2022-04-08 #64 @lbeziaud
@oliviaguest Tristan Allard asked [Julien Rossi](https://www.julienrossi.com/) for a recommendation: [Thomas Soubiran](https://pro.univ-lille.fr/thomas-soubiran/) might be interested
