### 2022-04-20 #64 @LukasWallrich
Also @oliviaguest you might need to make an editorial decision regarding how to label partial replications (or conceptual replications with minor discrepancies). I scanned #51 and they seem to report (in their initial submission) that they found quite a few discrepancies but generally concordant results as *[Re]* ... here that is proposed as *[-Re]* ... maybe there needs to be something in between?
