### 2022-04-21 #64 @rougier
Something like `[~Re]` or `[∂Re]`?
