### 2022-05-04 #64 @lbeziaud
@oliviaguest We wrote to Sean Reardon on 25/05/21 and to Rachel Baker on 28/05/21, asking if they would be comfortable sharing their code. As stated in the paper we received no answer. It would indeed be interesting to get their input if possible
