### 2022-05-16 #64 @oliviaguest
Indeed, very good idea and I was planning on it. What do we (editors and authors, @rougier @khinsen @benoit-girard and @lbeziaud) think about maybe asking the original authors to even review it (as a 3rd reviewer, not as the 2nd)?
