### 2022-06-20 #64 @lbeziaud
Thank you @LukasWallrich for the extended review. We updated the paper ([commit](https://github.com/lbeziaud/re-reardon2018/commit/ba86acf648575a6be91d1d2fe7abbf14f291702c) and [pdf diff](https://github.com/ReScience/submissions/files/8940803/diff.pdf) with additions in blue) and the code ([commit](https://github.com/lbeziaud/mosaic/commit/11e92b88de851ddfbfe7728331060e73ac5d0acf)) on the basis of your comments. We copy-paste them below and provide a concise answer to each.

> Unfortunately, I don't have suggestions for a second reviewer. However, I had a look at the manuscript and code- I enjoyed reading and reviewing it. In general, the code is largely clear and makes sense, while some bits of the paper (especially the results) definitely need to be clarified. Below, I list all my suggestions for improvements - if you don't like any of them, feel free to explain why that is the case. Also, pls obviously let me know if anything is unclear.
>
> ## Repo
>
> The README is very helpful. It might be good to include instructions for Windows (instead of source: venv\Scripts\activate), but not essential. I could install dependencies & appreciate that you added a licence

We are sorry but we do not have any Windows license to port the instructions.

> ## Code
>
> I can generally follow the code and can re-create the figures.
>
> ### Code
>
> More comments would be good - the ToDo comment should be checked and removed (if that difference matters and is unresolved, mention it in the paper)

I added comments and removed the TODO.

> Some variable names require guessing - e.g., expyield, sach & cach - I'd encourage you to use longer and clearer names, as recommended by most styleguides, but if you don't like that, comments can also help. Also, when you abbreviate statistical concepts (mean, err, etc.) it'd be great to always separate them (ach_mean rather than mach - unless I misunderstand mach)

I renamed the variables.

> ### Notebook
>
> Could you add sub-titles? The key benefit of that format over .py seems to be navigability - yet currently it is very hard to see what each block does. At least, please indicate which section replicates which of _your_ figures - that is the most likely thing a reader of your code would like to find out.

I restructured the notebook with sections. Most data and plot methods are now in `utils.py` and results are saved.

> ## Paper
>
> Add abstract - and clarify claim in the introduction: not able to replicate all. Do you think your results raise serious doubts about their conclusions?

Added abstract and clarified the introduction (no doubts raised about the conclusions of the original paper but the results highlight the general need to open experimental artifacts).

> Model description: "a percentage of the (best) students is admitted" - do they just admit the best ranked applicants, with the number of offers determined by the historical acceptance rate? If so, that could be clarified without making your text any longer.

Clarified, thank you.

> Pls clarify which authors you contacted - currently it reads like the authors of [3]. However, I expect that you also contacted the authors of [1] and think that this should be included.

Clarified: we contacted the first two authors of [1] (which are the first and last authors of [3]).

> ### Parameters
>
> 4 SES - why is the third bullet point quoted relevant? Could you clarify "no success"?

I removed the third bullet point. An experiment has been added (Appendix A, Fig 12) to illustrate "no success". This is not satisfactory as the result is a much smaller impact than the original model.

> How sensitive are your results to your assumption re college quality reliability? It does not seem implausible that this would be lower than own achievement reliability (though I would concur with your guess if we need to choose).

Added an experiment (Fig 13, Appendix A) highlighting the negligible impact of assuming a typo (max 0.9) or not (max 0.7).

> Quality update window: did 5 years not yield satisfying results (the first statement is - just about - compatible with updating a running average, so that would seem to be the assumption closer to the reporting)

Clarified. Using only the previous year (as per the model description) to update colleges' quality gives satisfying results. We assume the 5 years window is used to smooth some plots in the original paper.

> Initial college quality: clarify that 130 is the standard deviation (currently it sounds like it is the variance)

Clarified. It is the standard deviation. We switch to N(μ,σ²)

> Admission probability estimation: why does this need to refer to previous years? (obviously the original paper will say so, but a half-sentence here wouldn't hurt)

Clarified. Admission probability is estimated by students using a logistic regression on past admission results given the gap between student's self-assessed achievement and student's perception of college's quality.

> Could you include a table with all parameters as an appendix (also the ones clearly reported)? Given that the original paper did evidently not report the full specification, it would be helpful to have it here.

Added 3 tables along the paper: student distribution, constants and experimental variables. I ommitted the other parameters such as, e.g., the formulas of the weights, because displaying them would overload the tables without helping understand the model.

> ### Figures
>
> Purely a matter of taste - but might it be better to compare original results (a) with replication results (b) - that would also make the text flow more smoothly

Done.

> Also, could you align the formatting more closely (ok to keep color, but particularly the scale labels should be identical, and the axes ideally be aligned so that visual comparison is easier ... or you could use the original b/w formatting so that the same legend applies to both. If you keep color, please add a legend to the figure, they should stand alone without having to read the text)

Done.

> Figure 4: what is shown here? For enrolment at top-4 colleges, I would expect four lines. Probably this shows multiple runs - but could that be more prominent?

The original caption was more readable: there is only one run but there is one line per college. All 40 colleges are represented, but only 4 (darker lines) use affirmative action policies.

> Figure 6: * what is the logic of only basing this one on several runs? Also, why not show all the results? (Are the results means?).

Clarified. This is the only figure which requires averaging multiple runs (to reduce the noise).

> * More importantly, their left-most arrow is special - you need to explain that. Did you have any students that did not enrol? (Based on the model description, those would only be students who got no offers?) What did you do with them? Maybe you should add them to this figure.

Clarified (as a sentence, our figure is the same)

> * Caption: colors seem to be switched (orange is using), and the caption reads as if the colors show whether the top 4 schools use affirmative action - but that is not right, so please clarify

Done (black and white).

> Figure 7 - align figures for comparability (probably by splitting theirs into two rows like yours)

Done (to the best of my abilities without weird latex spaces and without manual cropping).

> Figures 8 & 9: their version with stacked bars appears a lot more comprehensible to me - so I would suggest sticking to that. Alternatively, please find another way where the reader does not need to compare small variations in bar height all the way across the page

Done.

> If you have code to recreate the full set of figures in the original paper, my tendency would be to include that into the Notebook - in case others want to build on your work (or just use a cc-licenced version of a figure)

The original figures are extracted (manually) from the pdf. I do not have the code to create them.

> ### Results
>
> I think your figure numbering changed at some point? At least I got confused - probably the first two comments below just require correct referenes?

Clarified reproduction vs original figure numbering.

> Figure 3: I don't understand your conclusion - it seems that your model shows _greater_ impact on minority students?

Nice spot, fixed.

> Figure 4: here the replication clearly differs from the original - at least more than in 3 - so I don't understand how the combination is satisfactory?

Clarified. Actually, the range (min, max) is ok, but the average is lower.

> Can you clarify the size of the differences between your and their results - for instance in terms of the shift in academic achievement z-scores based on SES required to match "real-world" race-based affirmative action? Something like that would seem to be needed to see whether your results should change the interpretation of their paper.

Without a clear understanding of their process a comparison is difficult...

> You conclude that SES affirmative action was weaker than expected - two potential reasons (you have likely thought about them, but then pls say so):  * What happens if you don't truncate the SES weights, and actually penalise high-SES students? I think your approach makes more sense, but could imagine that the original implementation did not do that, thus strengthening the effects they got?

In order to illustrate such impact, I have added Fig 13 (Appendix A). This results in a much higher positive impact on minorities than the original model.

> * Footnote 3: might this be a difference between your results and the original results? If so, could that explain the difference you observe in Figures 8 and 9?

This would result in incomparable categories but I did not try. The difference observed in Fig 8 and 9 are more likely due to an issue in understanding the policies in the original paper.

> ### Conclusion
>
> I would have expected you to conclude that open code is important? :)

Absolutely, done :)

> ### Other questions
>
> Using 40 colleges is quite a small sample - and the trajectories differ widely. In terms of assessing robustness, might that be a parameter worth experimenting with?

I am not sure that this would be worth. First, changing the number of colleges is not straightforward. Number of students, seats, applications are given with 40 colleges. Changing this parameter may require adjusting other "internal" parts of the model. Second, we believe that increasing the number of colleges would lead to similar trends because the observations on 40 colleges appear to be quite robust as we can observe from the figures where trends are rather identical for each colleges.

> In your figures 2-4, you have a notable divergence at the very start when the blue and orange lines move away from each other, while that is not evident in the original. Is that just a quirk of your random seed?

The first 15 years are a "burn-in period" (from original paper) which allows the model to stabilize. This includes the logistic regression _on the admission probabilities_. We could play with the initial coefficients of the LR to adjust but did not consider this variation an issue (since it's a boostraping phase).

> ### Minor language notes
>
> Section 2: completely -> complete Section 4, bottom: he -> the Initial college quality: relies -> relied

Fixed, thank you.
