### 2022-07-01 #64 @LukasWallrich
Thank you for making these changes. Things are a lot clearer now. Just a few more specific points.

# Code

I appreciate the tidy-up. However, it might be good to include one example for how to run an experiment into the notebook (or README if you prefer). Currently, one needs to dig quite deep into utils.py to figure that out, which makes it harder to reuse your code.

Also, it might be good to rename the file given that utils is a standard python library (I initially searched Google for the utils make_all() function :) )

# Paper

In 2 you say that affirmative action policies are introduced by some colleges, _mostly_ the top four. What does that mean? If some experiments involve different/more colleges, then that might need to be stated more clearly.

[Btw – I find it a bit strange that in any run, only one policy is adopted? An obvious question would be how policies interact – i.e. when different colleges focus on different policies in the same year, which is what will happen in reality. I am not suggesting that you now implement that here, but it might be worth pointing towards that in the need for further research.]

I understand that it is difficult to interpret the effect size / importance of the divergence – but I still think that should be addressed. Can you say what one would need to check to figure that out?

The difference in Figure 7 seems more important than other divergences. In your model, two of the top colleges do not appear to decline in quality at all. That contradicts a strong claim in the original paper: “As figure C5 makes clear, the top four ranked colleges, which begin using affirmative action in year 15, all experience a gradual decline in their quality ranking over time, as they continue to implement very aggressive affirmative action and recruitment strategies. It is unlikely that top colleges in the real world would so willingly sacrifice their quality ranking.” If you agree with my interpretation, that should be stated explicitly.

In the conclusion, you say that “more effort are required _regarding this replication_” – what does that mean?

# Details

-	Page 4 – something is going wrong in the sentence “as in the first quite would”
-	On p. 6 “probability admission” should be “admission probability”?
-	At the end “for fairness applications” does not work.
-	Be consistent whether you speak of “schools” or “colleges”
-	Your Figure 6 only has three black arrows. How do they represent 4 colleges?
-	In Figure 7, the order of the rows does not match – that confused me quite a lot when I tried to compare the results.
-	In Figure 10, the legends don’t match – and the distributions are strikingly different. Do you have the wrong chart from the original paper there?
