+++ utils.py
@@ -1,312 +1,0 @@
-from pathlib import Path
-from typing import Callable, Any, Union
-
-import numpy as np
-import pandas as pd
-import seaborn as sns
-from joblib import Parallel, delayed
-from matplotlib import pyplot as plt
-from scipy.stats import norm
-
-from reardon import RACES, run, RACE_WEIGHT, RES_LOC, RES_STD
-
-
-def run_agg_colls(num_repeats=10, **kwargs):
-    """
-    College dataframe as an average of multiple runs
-    """
-    runs_ = Parallel(n_jobs=-1, verbose=1)(
-        delayed(run)(**kwargs) for _ in range(num_repeats)
-    )
-    colls, _, _, _ = zip(*runs_)
-    # average of runs
-    colls = np.rec.fromarrays(
-        [
-            np.mean([colls_[name] for colls_ in colls], axis=0)
-            for name in colls[0].dtype.names
-        ],
-        dtype=colls[0].dtype,
-    )
-    # as dataframe
-    ind_yc = pd.MultiIndex.from_product(
-        (range(colls.shape[0]), range(colls.shape[1])), names=["year", "coll"]
-    )
-    colls = pd.DataFrame(colls.ravel(), index=ind_yc)
-    return colls
-
-
-def pct_enroll_by_race(**kwargs):
-    """
-    Percentage of enrolled students by race for active colleges after year 25
-    """
-    _, _, _, enr = run(**kwargs)
-    enr = enr[(enr.year >= 25) & enr.active]
-    races, counts = np.unique(enr.race, return_counts=True)
-    pct = counts * 100 / np.sum(counts)
-    names = np.take(RACES, races)
-    return pd.Series(pct, index=names)
-
-
-# mixture of resources distribution
-res_loc_ = np.sum(RACE_WEIGHT * RES_LOC)
-res_var = np.sum(RACE_WEIGHT * (RES_STD**2 + RES_LOC**2 - res_loc_**2))
-# common resource quintiles for all runs
-res_bins = norm.ppf(np.linspace(0, 1, 5 + 1), loc=res_loc_, scale=np.sqrt(res_var))
-
-
-def pct_enroll_by_res(**kwargs):
-    """
-    Percentage of enrolled students by resource quintile for active colleges after year 25
-    """
-    _, _, _, enr = run(**kwargs)
-    enr = enr[(enr.year >= 25) & enr.active]
-    counts, _ = np.histogram(enr.res, bins=res_bins)
-    pct = counts * 100 / np.sum(counts)
-    q = [f"Q{i}" for i in range(1, 5 + 1)]
-    return pd.Series(pct, index=q)
-
-
-def df_param_grid(param_grid):
-    return pd.MultiIndex.from_product(
-        param_grid.values(), names=param_grid.keys()
-    ).to_frame(index=False)
-
-
-def df_exp_grid(param_grid, exp):
-    df = df_param_grid(param_grid)
-    runs = Parallel(n_jobs=-1, verbose=1)(
-        delayed(exp)(**conf) for _, conf in df.iterrows()
-    )
-    df = df.join(pd.concat(runs, axis=1).T)
-    return df
-
-
-def df_melt_compo_by(df, var_name):
-    # id_vars = df.filter(like="wgt_").columns
-    # df = df.melt(
-    #     id_vars=id_vars,
-    #     var_name=var_name,
-    #     value_name="Composition (%)",
-    # )
-    df = df.rename(
-        columns={
-            "wgt_aa_race": "Race weight",
-            "wgt_aa_ses": "SES weight",
-            "wgt_recr_race": "Recruit weight",
-        }
-    )
-    return df
[…]
