+++ model.py
@@ -1,0 +1,279 @@
+from collections import namedtuple
+
+import numpy as np
+import pandas as pd
+from scipy.stats import zscore
+from sklearn.linear_model import LogisticRegression
+
+N_YEARS = 30
+
+N_COLLS = 40
+N_STUDS = 10000
+
+N_ACTIVE = 4
+
+N_SEATS = 150
+
+QUAL_LOC, QUAL_SCALE = 1070, 130
+
+WGT_AA_RACE = 0
+WGT_AA_SES = 0
+WGT_RECRUIT = 0
+
+RACE_NAMES = ["Asian", "Black", "Hispanic", "White"]
+
+RACE_WEIGHTS = np.array([0.05, 0.15, 0.2, 0.6])
+
+RES_LOCS = np.array([0.012, -0.224, -0.447, 0.198])
+RES_SCALES = np.array([0.833, 0.666, 0.691, 0.657])
+
+ACH_LOCS = np.array([1038, 869, 895, 1052])
+ACH_SCALES = np.array([202, 169, 185, 186])
+
+AR_CORRS = np.array([0.441, 0.305, 0.373, 0.395])
+
+
+S = namedtuple("S", ("colleges", "students", "outcomes"))
+
+
+def run(
+    *,
+    n_years=N_YEARS,
+    n_colls=N_COLLS,
+    n_studs=N_STUDS,
+    n_active=N_ACTIVE,
+    n_seats=N_SEATS,
+    qual_loc=QUAL_LOC,
+    qual_scale=QUAL_SCALE,
+    wgt_aa_race=WGT_AA_RACE,
+    wgt_aa_ses=WGT_AA_SES,
+    wgt_recruit=WGT_RECRUIT,
+    race_weights=RACE_WEIGHTS,
+    res_locs=RES_LOCS,
+    res_scales=RES_SCALES,
+    ach_locs=ACH_LOCS,
+    ach_scales=ACH_SCALES,
+    ar_corrs=AR_CORRS,
+    seed=None,
+    ses_alt=False,
+    ses_penalty=False,
+    obs_qual_alt=False,
+):
+    """
+
+    :param n_years: number of years to simulate
+    :param n_colls: number of colleges
+    :param n_studs: number of students
+    :param n_active: number of colleges implementing policies
+    :param n_seats: number of seats per colleges
+    :param qual_loc: mean of colleges' quality normal distribution
+    :param qual_scale: scale of colleges' quality normal distribution
+    :param wgt_aa_race: weight of race-based afformative action
+    :param wgt_aa_ses: weight of SES-based affirmative action
+    :param wgt_recruit: weight of race-based recruiting
+    :param race_weights: weights of racial distribution
+    :param res_locs: means of students' resources normal distribution
+    :param res_scales: scales of students' resources normal distribution
+    :param ach_locs: means of students' achievement normal distribution
+    :param ach_scales: scales of students' achievement normal distribution
+    :param ar_corrs: correlations between students' resources and achievement
+    :param seed:
+    :param ses_alt: alternative SES-based AA
+    :param ses_penalty: alternative SES-based AA
+    :param obs_qual_alt: alternative perception quality upper-bound
+    :return: namedtuple of dataframes with results for colleges, students and decisions
+    """
+    # weighted mixture
+    ach_loc_mix = np.sum(race_weights * ach_locs)
+    ach_var_mix = np.sum(
+        race_weights * (ach_scales**2 + ach_locs**2 - ach_loc_mix**2)
+    )
+
+    rng = np.random.default_rng(seed)
+
+    qual = np.empty((n_years, n_colls))
+    qual[0] = rng.normal(qual_loc, qual_scale, size=n_colls)
+    qual[0, ::-1].sort()  # sort by descending quality
+
+    # mask true if college (top n_active) will use policy
[…]
