+++ README.md
@@ -8,13 +8,13 @@ See https://github.com/lbeziaud/re-reardon2018 for details on the replication.
 
 ## Content
 
-The main model is [reardon.py](reardon.py). Figures are generated by [figures.ipynb](figures.ipynb) (with some parallel processing).
+The main model is [model.py](model.py). Figures are generated by [figures.ipynb](figures.ipynb). The file [utils_figs.py](utils_figs.py) contains auxiliary functions used for launching the experiments. 
 
 ## Requirements
 
 This code has been written with Python 3.8. Required packages are listed in [requirements.txt](requirements.txt)
 
-## Installation
+## Usage
 
 1. Clone the repository
 
@@ -26,7 +26,7 @@ $ cd mosaic
 2. Install dependencies (in a virtualenv)
 
 ```console
-$ python3.9 -m venv venv  # create venv
+$ python3.8 -m venv venv  # create venv
 $ source venv/bin/activate  # activate venv
 (venv) $ pip install -r requirements.txt
 ```
@@ -36,3 +36,18 @@ $ source venv/bin/activate  # activate venv
 ```console
 (venv) $ jupyter nbconvert --ExecutePreprocessor.timeout=-1 --to notebook --inplace --execute figures.ipynb
 ```
+
+4. Running the model is done by calling `model.run`, which returns a tuple of dataframes:
+
+```python
+from model import run
+
+colleges, students, outcomes = run()
+
+colleges.to_csv("colleges.csv")
+students.to_csv("students.csv")
+outcomes.to_csv("outcomes.csv")
+# See also https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html
+```
+
+Several examples are provided in [utils_figs.py](utils_figs.py) on how to cache the run and on how to run and aggregate repeated experiments.

[…]
