### 2022-08-08 #64 @lbeziaud
Sorry for the delay and thank you again @LukasWallrich for your feedback. I updated the paper ([commit](https://github.com/lbeziaud/re-reardon2018/commit/6acdc5f54e6b23f996e3f178d0a31661893f4bef) and [pdf diff](https://github.com/ReScience/submissions/files/9283111/diff.pdf)) and the code ([commit](https://github.com/lbeziaud/mosaic/commit/4827206b77fe04e51c19e32839f5d41ec401bceb)).

> Thank you for making these changes. Things are a lot clearer now. Just a few more specific points.

> # Code
>
> I appreciate the tidy-up. However, it might be good to include one example for how to run an experiment into the notebook (or README if you prefer). Currently, one needs to dig quite deep into utils.py to figure that out, which makes it harder to reuse your code.

The README has been updated to include additional info, with some code cleanup to (hopefully) make running experiments easier.

> Also, it might be good to rename the file given that utils is a standard python library (I initially searched Google for the utils make_all() function :) )

I could not find utils in the standard library (tested with `python -I -m utils`) but renamed the file to "utils_figs.py" :)

> # Paper
>
> In 2 you say that affirmative action policies are introduced by some colleges, _mostly_ the top four. What does that mean? If some experiments involve different/more colleges, then that might need to be stated more clearly.

Removed "mostly the top four". The original paper present additional experiments with 10, 20 and 40 active colleges which I added in appendix.

> [Btw – I find it a bit strange that in any run, only one policy is adopted? An obvious question would be how policies interact – i.e. when different colleges focus on different policies in the same year, which is what will happen in reality. I am not suggesting that you now implement that here, but it might be worth pointing towards that in the need for further research.]

This is actually close to a question we ask ourselves using this model, in the context of algorithmical fairness.

> I understand that it is difficult to interpret the effect size / importance of the divergence – but I still think that should be addressed. Can you say what one would need to check to figure that out?

The obvious approach is to try (again) to access the original code, the alternative being trying to identify the actual source(s) of our misunderstanding. We are not domain experts but to interpret the impact I would implement your aforementioned suggestion of mixing policies. Having both colleges obeying our interpretation and the original would highlight the actual divergence in terms of their impact on students, racial and SES groups, and colleges.
That being said, I think the divergence results from technical misunderstandings, which might not warrant a more in depth interpretation (but rather a clarification).

> The difference in Figure 7 seems more important than other divergences. In your model, two of the top colleges do not appear to decline in quality at all. That contradicts a strong claim in the original paper: “As figure C5 makes clear, the top four ranked colleges, which begin using affirmative action in year 15, all experience a gradual decline in their quality ranking over time, as they continue to implement very aggressive affirmative action and recruitment strategies. It is unlikely that top colleges in the real world would so willingly sacrifice their quality ranking.” If you agree with my interpretation, that should be stated explicitly.

This is indeed an important distinction between our results. Figures in previous submissions were the result of only one run. I updated all of them to an average of 10 runs (as the original paper). The observation is still valid. I assume this is the result of our lower impact of the policies, which allows colleges to not admit students with the same low achievement as the original work. I added a reference to the strong claim in our results section.

> In the conclusion, you say that “more effort are required _regarding this replication_” – what does that mean?

Reworded.

> # Details
>
>     * Page 4 – something is going wrong in the sentence “as in the first quite would”
>
>     * On p. 6 “probability admission” should be “admission probability”?
>
>     * At the end “for fairness applications” does not work.
>
>     * Be consistent whether you speak of “schools” or “colleges”

Fixed, thank you.

>     * Your Figure 6 only has three black arrows. How do they represent 4 colleges?

Fixed. The rightmost arrow was outside the boundary.

>     * In Figure 7, the order of the rows does not match – that confused me quite a lot when I tried to compare the results.

Fixed.

>     * In Figure 10, the legends don’t match – and the distributions are strikingly different. Do you have the wrong chart from the original paper there?

Fixed. It was indeed the wrong chart 🙃. Thank you.

**from your previous comments:**

> More importantly, their left-most arrow is special - you need to explain that. Did you have any students that did not enrol? (Based on the model description, those would only be students who got no offers?) What did you do with them? Maybe you should add them to this figure.

"The left-most arrow captures students who do not enroll in college in our simulation." Figures have been fixed to show this arrow, noting however that we miss some context regarding this arrow (are unenrolled students admitted students, students having applied, etc.).
