### 2022-08-12 #64 @LukasWallrich
Thanks for making these updates - I think this all looks good now and I appreciate the clarifications in the paper. Tiny point: on top of p.6 (in the PDF diff) there is a broken Appendix reference (??). @oliviaguest please let me know if you need anything else from me.
