### 2022-08-13 #64 @lbeziaud
> Tiny point: on top of p.6 (in the PDF diff) there is a broken Appendix reference (??).

Thank you for catching that. The broken reference is only present in the diff, and not in the article :) I'm still discovering [latexdiff](https://ctan.org/pkg/latexdiff).

@LukasWallrich Thank you again for your great feedback. I think the paper and code gained a lot from it!
