+++ requirements.txt
@@ -1,8 +1,8 @@
-joblib==1.1.0
-jupyter
-matplotlib==3.5.1
-numpy==1.22.3
-pandas==1.4.1
-scikit-learn==1.0.2
-scipy==1.8.0
-seaborn==0.11.2
+joblib==1.2.0
+matplotlib==3.6.2
+notebook==6.5.2
+numpy==1.23.4
+pandas==1.5.1
+scikit-learn==1.1.3
+scipy==1.9.3
+seaborn==0.12.1
