+++ README.md
@@ -12,7 +12,7 @@ The main model is [model.py](model.py). Figures are generated by [figures.ipynb]
 
 ## Requirements
 
-This code has been written with Python 3.8. Required packages are listed in [requirements.txt](requirements.txt)
+This code has been written with Python 3.9. Required packages are listed in [requirements.txt](requirements.txt)
 
 ## Usage
 
@@ -26,7 +26,7 @@ $ cd mosaic
 2. Install dependencies (in a virtualenv)
 
 ```console
-$ python3.8 -m venv venv  # create venv
+$ python3.9 -m venv venv  # create venv
 $ source venv/bin/activate  # activate venv
 (venv) $ pip install -r requirements.txt
 ```
@@ -43,11 +43,6 @@ $ source venv/bin/activate  # activate venv
 from model import run
 
 colleges, students, outcomes = run()
-
-colleges.to_csv("colleges.csv")
-students.to_csv("students.csv")
-outcomes.to_csv("outcomes.csv")
-# See also https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html
 ```
 
 Several examples are provided in [utils_figs.py](utils_figs.py) on how to cache the run and on how to run and aggregate repeated experiments.
