### 2022-11-07 e60e7d3
update readme and notebook
