### 2022-11-07 #64 @lbeziaud
Hello, I switched to python 3.9 and updated the dependencies due to a [security issue with joblib < 1.2.0](https://github.com/advisories/GHSA-6hrg-qmvc-2xh8). The code hasn't changed.

Please tell me if anything needs to be done on our side now :)
