### 2023-04-29 #64 @cosimameyer
Hi everyone,

Thanks so much for your patience and for inviting me to review your paper. I enjoyed reading it and going through your code 😊

I think this journal (and the contributions) are extremely valuable (but yet often undervalued), so thanks to everyone involved in the process of running it!

I'm listing my feedback and suggestions below. I reviewed [the paper and the code listed here](https://github.com/ReScience/submissions/issues/64#issue-1172406583) as the latest commit dates indicate that there were changes made after the first review. If you don't agree with them, feel free to explain why and please also let me know if anything is unclear.

## Repository

### Code

The guide in the README was easily accessible and I could execute the code successfully.
However, I ran into two challenges: 1) the storage capacity, and 3) the execution time of the script. I would add in the README that you need a minimum of ca. 40 GB of storage capacity to run the script because the models are cached and also add your system specifications. This might help future users to get a reference point. For instance, it took me around 1.75 hours to convert and to execute `figures.ipynb` (I am working on a MacBook Pro with a 2 GHz Quad-Core Intel Core i5 and 16 GB RAM).

#### Code
- The code is complete and executable. The description is sufficient to run the code successfully. At times, I would have wished for a bit more documentation. For instance, the docstring in `plot_d123` only says `"""Plot figures D1, D2, D3."""` but doesn't provide more info - but this may be also a matter of taste.

#### Notebook
- I found it at times a bit confusing which figures relate to the respective figure in the paper. But as a general recommendation, I would name/enumerate the figures consistently and separate the notebook into "Figures in the main paper" and "Figures in the appendix" (or something similar).
- I am not sure why but I think Figure 3 (see first screenshot) differs significantly from the figure provided as a replication in your paper (see second screenshot).

Figure 3 as generated using the code provided (including code to understand which chunk generated the figure):

Figure 3 in the paper:

## Paper

The paper is easily readable and informative. I listed some suggestions for more clarification below.

#### Introduction
- You mention in the beginning of your paper on page 1, that "while others were used inconsistently along the paper–such as affirmative actions." Could you specify in another short sentence (with an example) how they were used inconsistently?

#### Model of college admission
- I would add an additional sentence making clear that the data are fully simulated - both in the original paper and in your replication. This may help a reader less familiar with agent-based simulation models to understand where the data come from.
- If possible, it is good to add a comparison of the distribution of the replication and the original data to better understand whether they are comparable.

#### Method
- You mention that you used `sklearn` but also relied on `statsmodel` (during development) - how do the results differ? Why did you lean towards `sklearn` in the end?

#### Results
- Figure 2: While I agree that there's success in the replication, it is interesting to see that there seem to be reverse trends in the beginning ($\leq$ 3 years)
- Figure 6: You mention that "[i]t is unspecified however whether it captures students who do not enroll while having applied, being admitted or without condition. We use the latter." (p. 6) - what happens if you use the former? Do the results change?
- Figure 7: It is not easy to compare Figure 7 with the original because of the different structure. Would it be possible to arrange it similarly to the original paper?
- On a general note: To be better able to compare both the replication and the original results, it might be worth the effort to add similar scales (and ticks)  on the axes and to place the figures on a similar level. I know this can be tricky but it'll definitely help the reader to easily spot differences (and similarities).

#### Additional Figures
- I like that you add additional figures but couldn't find references to them in the text (except Figure 14 which is mentioned in Section 4). I think it might be good to spend words describing the figures (and what their interpretation means) and link to them in the text (as done with Figure 14)

## General remarks
- It is mentioned in the comments above but since the paper cannot fully replicate the original study, I'm definitely supporting reaching out to the authors (once more and now also backed by the paper and the reviews) and getting their stance on the "unspecified and inconsistent parameters". I think with what you had at hand, it's still impressive to see how much of the original study you were able to replicate!

- Just a general thought which doesn't have to be addressed anywhere: [Reardon et al. (2018) used Stata in their original paper](https://www.comses.net/codebases/4220/releases/1.0.0/) and it would be interesting to know whether the programming language used (and essentially the way the algorithms are written) could have had also an effect on differing results.

## Small remarks

- "[...], which leads us to believe" (instead of "lead" on page 6 in the paragraph where you describe Figure 4)
