+++ utils_figs.py
@@ -1,10 +1,10 @@
 from itertools import product, starmap
+from pathlib import Path
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import seaborn as sns
-from joblib import Memory
 from scipy.stats import norm
 
 from model import RACE_WEIGHTS, RES_LOCS, RES_SCALES
@@ -12,16 +12,51 @@ from model import run as run_uncached
 
 sns.set_style("whitegrid")
 
-# setup cached model run
-location = "./cache"
-memory = Memory(location, verbose=0, mmap_mode="r")
-run = memory.cache(run_uncached)
+location = Path("./results")
+
+DEFAULT = {
+    "n_years": 30,
+    "n_active": 4,
+    "wgt_aa_race": 0,
+    "wgt_aa_ses": 0,
+    "wgt_recruit": 0,
+    "obs_qual_alt": False,
+    "ses_penalty": False,
+    "ses_alt": False,
+    "seed": None
+}
+
+
+def kwargs_default(**kwargs):
+    return DEFAULT | kwargs
+
+
+def run(**kwargs):
+    kwargs = kwargs_default(**kwargs)
+    pattern = "/".join(f"{k}~{v}" for k, v in kwargs.items())
+    dirname = location / pattern
+
+    try:
+        colleges = pd.read_parquet(dirname / "colleges.parquet")
+        students = pd.read_parquet(dirname / "students.parquet")
+        outcomes = pd.read_parquet(dirname / "outcomes.parquet")
+    except FileNotFoundError:
+        colleges, students, outcomes = run_uncached(**kwargs)
+
+        dirname.mkdir(parents=True, exist_ok=True)
+
+        colleges.to_parquet(dirname / "colleges.parquet")
+        students.to_parquet(dirname / "students.parquet")
+        outcomes.to_parquet(dirname / "outcomes.parquet")
+
+    return colleges, students, outcomes
+
 
 # number of time each run is repeated
-n_repeats = 10
+N_REPEATS = 10
 
 # initial seed (use to generate subsequent seeds)
-entropy = 286130738418515439860296005820123389689
+ENTROPY = 286130738418515439860296005820123389689
 
 # number of quantile to consider for students' resources
 n_res_bins = 5
@@ -41,6 +76,12 @@ order_race = ["Black", "Hispanic", "Asian", "White"]
 order_res = ["Q1", "Q2", "Q3", "Q4", "Q5"]
 
 
+def kwargs_default_seeded(**kwargs):
+    sq = np.random.SeedSequence(ENTROPY)
+    seeds = sq.generate_state(N_REPEATS)
+    yield from (kwargs_default(**kwargs) | dict(seed=seed) for seed in seeds)
+
+
 def dict_product(d):
     """Cartesian product of input iterables with keys."""
     # dict_product({"A": [1, 2], "B": [3]}) --> [{'A': 1, 'B': 3}, {'A': 2, 'B': 3}]
@@ -49,9 +90,7 @@ def dict_product(d):
 
 def run_repeat(**params):
     """Run (cached) model multiple times with distinct seeds. Initial seed is constant."""
-    sq = np.random.SeedSequence(entropy)
-    seeds = sq.generate_state(n_repeats)
-    return (run(**params, seed=seed) for seed in seeds)
+    return (run(**params) for params in kwargs_default_seeded(**params))
 
 
 def run_repeat_colleges(**params):
@@ -80,13 +119,19 @@ def grid_run(param_grid, func):
     )
 
 
-def pct_race_enroll(colleges, students, outcomes):
-    """Percentage of total enrolled students after year 25 in active colleges by race."""
+def studs_enrolled_active_after25(colleges, students, outcomes):
     return (
[…]
