+++ requirements.txt
@@ -3,6 +3,7 @@ matplotlib==3.6.2
 notebook==6.5.2
 numpy==1.23.4
 pandas==1.5.1
+pyarrow==11.0.0
 scikit-learn==1.1.3
 scipy==1.9.3
 seaborn==0.12.1
