+++ README.md
@@ -8,7 +8,7 @@ See https://github.com/lbeziaud/re-reardon2018 for details on the replication.
 
 ## Content
 
-The main model is [model.py](model.py). Figures are generated by [figures.ipynb](figures.ipynb). The file [utils_figs.py](utils_figs.py) contains auxiliary functions used for launching the experiments. 
+The main model is [model.py](model.py). Figures are generated by [figures.ipynb](figures.ipynb). The file [utils_figs.py](utils_figs.py) contains auxiliary functions used for launching the experiments, with [pre_run](pre_run.py) allowing to run all experiments in parallel. 
 
 ## Requirements
 
@@ -37,9 +37,21 @@ This code has been written with Python 3.9. Required packages are listed in [req
     (venv) $ jupyter nbconvert --ExecutePreprocessor.timeout=-1 --to notebook --inplace --execute figures.ipynb
     ```
    
-   The whole notebook takes around 30 minutes to execute. Note that execution is cached on disk.
+   The whole notebook takes around 20 minutes to execute if the simulation files are already available (see below).
+   
+4. You can run all simulations required by the figures with:
 
-4. Running the model is done by calling `model.run`, which returns a tuple of dataframes:
+    ```console
+    (venv) $ python pre_run.py
+    ```
+    
+    The script accepts parameters `-j <number of jobs to run in parallel>` and `-v/-vv/-vvv` to control verbosity.
+    
+    Output is saved on disk in `results/` (in human readable paths) as parquet files. Total size is 10 GB.
+    
+    Using 15 parallel jobs requires around 30 GB of memory and takes 20 minutes on AMD Ryzen 7 PRO 6850U. 
+
+5. Running the model is done by calling `model.run`, which returns a tuple of dataframes:
 
     ```python
     from model import run
