### 2023-05-17 8eaabdc
fix doc plot_d123; fix lint warnings
