### 2023-05-23 #64 @lbeziaud
Thank you for your detailed feedback! I've updated the code and paper. Below is a quick answer to every point.

> ## Repository
> ### Code
>
> The guide in the README was easily accessible and I could execute the code successfully. However, I ran into two
> challenges: 1) the storage capacity, and 3) the execution time of the script. I would add in the README that you need a
> minimum of ca. 40 GB of storage capacity to run the script because the models are cached and also add your system
> specifications. This might help future users to get a reference point. For instance, it took me around 1.75 hours to convert
> and to execute `figures.ipynb` (I am working on a MacBook Pro with a 2 GHz Quad-Core Intel Core i5 and 16 GB RAM).

Sorry for that!
I improved the caching with manual storage as parquet in a human-readable path instead of using joblib.
The cache now takes ~ 10 GB.
I have also added a script `pre_run.py` which can create all required data with parallel processing.
I have added an indication of the disk and time requirements in the README.
On my laptop (AMD Ryzen 7 PRO 6850U) generating the data with 15 threads takes 30 GB of memory and 20 minutes.
Plotting the figures stills requires some processing and takes 20 minutes (once the data is generated).

> #### Code
>
> * The code is complete and executable. The description is sufficient to run the code successfully. At times, I would have
    wished for a bit more documentation. For instance, the docstring in `plot_d123` only
    says `"""Plot figures D1, D2, D3."""` but doesn't provide more info - but this may be also a matter of taste.

The docstring for `plot_d123` is now a bit more explicit.

> #### Notebook
>
> * I found it at times a bit confusing which figures relate to the respective figure in the paper. But as a general
    recommendation, I would name/enumerate the figures consistently and separate the notebook into "Figures in the main
    paper" and "Figures in the appendix" (or something similar).

I fixed the numbering to indicate when figures are in the main paper of in the appendix. Instead of `Figure [X]` everywhere (which was a mistake), figures are now numbered as they are in our paper, with the appendix letter prefixed when appropriate. I kept the same order since this way figures are grouped by type, but I can split it as main/appendix if you prefer?

> * I am not sure why but I think Figure 3 (see first screenshot) differs significantly from the figure provided as a
    replication in your paper (see second screenshot).
>
>
> Figure 3 as generated using the code provided (including code to understand which chunk generated the
> figure): <img alt="Screenshot 2023-04-29 at 16 21 03" width="643" src="https://user-images.githubusercontent.com/7782895/235308988-09cb47d5-a491-4320-9184-45dfe987e966.png">
>
> Figure 3 in the
> paper: <img alt="image" width="947" src="https://user-images.githubusercontent.com/7782895/235308946-147ade2c-276e-40d5-a555-ed04adfbca5e.png">

This is related to you comment about fixing the numbering. Figure 3 in the paper is actually the second figure in the
notebook (named C2 in the notebook and original paper), whereas the third figure in the notebook (named C3) is Figure 15 in
Appendix B.
Your first screenshot (notebook) is Figure 15 (C3) and not Figure 3 (C2).
This was indeed confusing, sorry. I hope fixing the numbering in the notebook also fixes the confusion?

> ## Paper
>
> The paper is easily readable and informative. I listed some suggestions for more clarification below.
> #### Introduction
>
> * You mention in the beginning of your paper on page 1, that "while others were used inconsistently along the paper–such as
    affirmative actions." Could you specify in another short sentence (with an example) how they were used inconsistently?

I added [an example](https://github.com/lbeziaud/re-reardon2018/blob/81f3bf1cad5412c9ca2d8339c9227f298329add3/content.tex#L14)

> #### Model of college admission
>
> * I would add an additional sentence making clear that the data are fully simulated - both in the original paper and in
    your replication. This may help a reader less familiar with agent-based simulation models to understand where the data
    come from.

I added [a sentence](https://github.com/lbeziaud/re-reardon2018/blob/81f3bf1cad5412c9ca2d8339c9227f298329add3/content.tex#LL23C6)

> * If possible, it is good to add a comparison of the distribution of the replication and the original data to better
    understand whether they are comparable.

Without access to the original data I am not certain how compare the distributions. Am I misunderstanding your comment?

> #### Method
>
> * You mention that you used `sklearn` but also relied on `statsmodel` (during development) - how do the results differ? Why
    did you lean towards `sklearn` in the end?

I added [a footnote](https://github.com/lbeziaud/re-reardon2018/blob/81f3bf1cad5412c9ca2d8339c9227f298329add3/content.tex#L139)

> #### Results
>
> * Figure 2: While I agree that there's success in the replication, it is interesting to see that there seem to be reverse
    trends in the beginning (≤ 3 years)

I added [the comment and a potential explaination](https://github.com/lbeziaud/re-reardon2018/blob/81f3bf1cad5412c9ca2d8339c9227f298329add3/content.tex#L376)

> * Figure 6: You mention that "[i]t is unspecified however whether it captures students who do not enroll while having
    applied, being admitted or without condition. We use the latter." (p. 6) - what happens if you use the former? Do the
    results change?

Looking only at applications gives similar results to looking at all students.
Looking only at admission gives values bigger than the original figure.
I am not certain it is an interesting experiment but I got the following data if you think I should add it to the paper?
It uses the same data as Figure 6 (left side but right side is identical), with the original having both arrows at ach=~800, %res=~0.6)

| filter | ach         | %race    | %res     |
|--------|-------------|----------|----------|
| any    | 990.010698  | 0.353296 | 0.598398 |
| app    | 979.329194  | 0.358160 | 0.591767 |
| adm    | 1177.594869 | 0.167540 | 0.397116 |

> * Figure 7: It is not easy to compare Figure 7 with the original because of the different structure. Would it be possible
    to arrange it similarly to the original paper?

I edited the original figures to combine them (they were splitted accross two pages) and played with the size to align them
better (with a focus on aligning y=1).

> * On a general note: To be better able to compare both the replication and the original results, it might be worth the
    effort to add similar scales (and ticks)  on the axes and to place the figures on a similar level. I know this can be
    tricky but it'll definitely help the reader to easily spot differences (and similarities).

I've tried to do that with "simple" customization available through matplotlib and seaborn but this is really tricky. Since I
only have the data for some of the original figures (8, 9, 10, 11) I cannot reproduce them with the same libraries I used, it
would have been a lot simpler than reproducing the look of Stata plots…

> #### Additional Figures
>
> * I like that you add additional figures but couldn't find references to them in the text (except Figure 14 which is
    mentioned in Section 4). I think it might be good to spend words describing the figures (and what their interpretation
    means) and link to them in the text (as done with Figure 14)

I've added references to figures of the appendix accross the Result section. Figures 12 and 13 are already referenced in the Difficulties section (SES‐based affirmative action).

> ## General remarks
>
> * It is mentioned in the comments above but since the paper cannot fully replicate the original study, I'm definitely
    supporting reaching out to the authors (once more and now also backed by the paper and the reviews) and getting their
    stance on the "unspecified and inconsistent parameters". I think with what you had at hand, it's still impressive to see
    how much of the original study you were able to replicate!

> * Just a general thought which doesn't have to be addressed
    anywhere: [Reardon et al. (2018) used Stata in their original paper](https://www.comses.net/codebases/4220/releases/1.0.0/)
    and it would be interesting to know whether the programming language used (and essentially the way the algorithms are
    written) could have had also an effect on differing results.

That is an interesting question and something I thought about for the logistic regression part. The Stata documentation gives
few details compared to the parameters offered by scikit-learn. Since the impact of the logit parameters was limited (mainly the first years, used as bootstrap only), I focused on speed and have not investigated further the replication side.

> ## Small remarks
>
> * "[...], which leads us to believe" (instead of "lead" on page 6 in the paragraph where you describe Figure 4)

Thank you for catching that!
