### 2023-06-22 #64 @cosimameyer
Thanks so much for your patience, getting back to this took longer than expected. Thanks again for carefully answering my questions - I highly appreciate it.

Please find my replies to open questions below - I'm happy to recommend the replication for publication :)

> I fixed the numbering to indicate when figures are in the main paper of in the appendix. Instead of Figure [X] everywhere (which was a mistake), figures are now numbered as they are in our paper, with the appendix letter prefixed when appropriate. I kept the same order since this way figures are grouped by type, but I can split it as main/appendix if you prefer?

I'd personally prefer a split in "main" and "appendix" (it makes logically more sense for me) but will leave the decision for you/the editor - I also see reasons why the current order makes sense.

 > This is related to you comment about fixing the numbering. Figure 3 in the paper is actually the second figure in the
notebook (named C2 in the notebook and original paper), whereas the third figure in the notebook (named C3) is Figure 15 in
Appendix B.
Your first screenshot (notebook) is Figure 15 (C3) and not Figure 3 (C2).
This was indeed confusing, sorry. I hope fixing the numbering in the notebook also fixes the confusion?

It does, thanks!

> Looking only at applications gives similar results to looking at all students.
Looking only at admission gives values bigger than the original figure.
I am not certain it is an interesting experiment but I got the following data if you think I should add it to the paper?
It uses the same data as Figure 6 (left side but right side is identical), with the original having both arrows at ach=~800, %res=~0.6)

Thanks for taking the effort - I think it would be a great fit for the appendix!
