### 2023-06-23 #64 @lbeziaud
Thank you @cosimameyer for the answer :)
I modified the notebook to [split between main and appendices](https://github.com/lbeziaud/mosaic/commit/d0dad3769fe556890dc454574cb82e24e5bd0026). I also added the [code used to produce the table of my previous reply](https://github.com/lbeziaud/mosaic/commit/328fd81ca58d10843642e08f900bba86d4675f16) and [an appendix (briefly) presenting those results](https://github.com/lbeziaud/re-reardon2018/commit/a85a8e130e4fb317fd3b83786f2ddb1fbd916238). (Beside that, I added an [acknowledgment section](https://github.com/lbeziaud/re-reardon2018/commit/19ee9cc87c4fbe1854b7c3fb8e40bac44dc0a9c0))

@rougier @oliviaguest Should I modify the title to [~Re] or [∂Re] as you discussed?
