### 2023-07-17 #64 @rougier
@lbeziaud @oliviaguest Sorry for delay in answer. Since I did not follow the submission from start to end, could you remind me what's the status of the replication (full, partial or negative) ?
