### 2023-07-17 #64 @lbeziaud
@rougier I would go with partial rather than negative. Most of the model is successfully replicated, but we are not successful regarding the actual affirmative action part (which is the subject of the paper). However, the trends are similar and the conclusions are not invalidated.
