### 2023-07-19 #64 @LukasWallrich
I think a partial replication label would be best - if you are willing to
introduce that. For instance, [~Re] - neither of the others really captures
the situation here.
