### 2023-07-19 #64 @rougier
@khinsen @benoit-girard Do you know/remember whihc notation we use for partial replication?
@LukasWallrich Can you tell me where are your latex and metadata files?
