### 2023-07-20 #64 @khinsen
@rougier There are various references to partial replications in our submission template, our Web site, and in discussion threads, but I can't find (nor remember) any decision about notation.
