### 2023-07-24 d0753ea
follow paper def bonus without race std in action  #3 (#4)
* add race ach std factor to sach cach #3

* no race ach std when active

* results without std when action #3
