+++ requirements.txt
@@ -5,5 +5,5 @@ numpy==1.23.4
 pandas==1.5.1
 pyarrow==11.0.0
 scikit-learn==1.1.3
-scipy==1.9.3
+scipy==1.10.0
 seaborn==0.12.1
