### 2023-08-16 #64 @lbeziaud
@oliviaguest @rougier In the meantime I've made a snapshot on Software Heritage https://archive.softwareheritage.org/swh:1:snp:bebbfb6e968909770c153e43232c0d0272c458f9;origin=https://github.com/lbeziaud/mosaic

edit: (09/09)
- I fixed two typos in the paper https://github.com/lbeziaud/re-reardon2018/commit/c3c85d967e0b3ad6ba8b9ee6f054efd5e18277d1
- switched the figures to pdf rather than png https://github.com/lbeziaud/mosaic/commit/4691a01b439094fc2460717d1a6944ce80c27d09 https://github.com/lbeziaud/re-reardon2018/commit/8b715a0396e12a08c23e8baecdaac9049ded2c14
- and updated the metadata https://github.com/lbeziaud/re-reardon2018/commit/6c7bc509af754309ff1fb7342e364f9a93298820
