### 2023-09-11 #64 @benoit-girard
I don't think we made any decision about partial replications, but I am fine with the [~Re] proposal.
