### 2023-10-02 #64 @lbeziaud
Hi @rougier @oliviaguest Is there anything left do do on my part?  I'm currently finalizing my thesis and would love to properly cite this work :)
