### 2023-10-13 #64 @rougier
Sorry for late answer. Can give me URL to your latex sources repository, that would be easier for publishing the article.
