### 2023-10-13 #64 @lbeziaud
@rougier the latex source is at https://github.com/lbeziaud/re-reardon2018/
