### 2023-11-16 #64 @rougier
Sorry for looooonng delay. I've tried to publish today but the Zenodo has changed and I need their input in order to modify the publish script. Sorry for this supplementary delay. Can you ping me here on next Monday to make sure it is not further delayed?
