### 2023-11-22 #64 @rougier
@oliviaguest  @lbeziaud Here is the sandbox version (to be checked before actual publication): https://sandbox.zenodo.org/record/773
