### 2023-11-23 #64 @lbeziaud
@rougier Thank you! Zenodo shows `\textasciitilde` rather than `~` in the title. Can we fix that or should I change something in the metadata?
