### 2023-12-04 #64 @rougier
It's online (for  real) at https://zenodo.org/record/10255347. It will soon appear on rescience website as well.
